requirements.txt
bottle==0.12.25
cffi==1.15.1
clr-loader==0.2.5
proxy-tools==0.1.0
pycparser==2.21
pythonnet==3.0.1
pywebview==4.0.2
